--- x/avs/keeper/epoch.go
@@ -1,12 +1,0 @@
-package keeper
-
-import (
-	abci "github.com/cometbft/cometbft/abci/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// EndBlock
-func (k *Keeper) EndBlock(ctx sdk.Context, _ abci.RequestEndBlock) []abci.ValidatorUpdate {
-
-	return []abci.ValidatorUpdate{}
-}
